% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osquery-driver.R
\docType{methods}
\name{dbDataType,OsqueryConnection-method}
\alias{dbDataType,OsqueryConnection-method}
\title{Find the database data type associated with an R object}
\usage{
\S4method{dbDataType}{OsqueryConnection}(dbObj, obj, ...)
}
\description{
Find the database data type associated with an R object
}
