% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osquery-driver.R
\docType{methods}
\name{dbFetch,OsqueryResult-method}
\alias{dbFetch,OsqueryResult-method}
\title{Retrieve records from Osquery query}
\usage{
\S4method{dbFetch}{OsqueryResult}(res, n = -1, ...)
}
\description{
Retrieve records from Osquery query
}
