% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr-helpers.R
\name{dbplyr_helpers}
\alias{dbplyr_helpers}
\alias{db_query_fields.OsqueryConnection}
\alias{db_data_type.OsqueryConnection}
\alias{sql_escape_ident.OsqueryConnection}
\alias{db_desc.OsqueryConnection}
\title{dbplyr_helpers}
\usage{
db_query_fields.OsqueryConnection(con, sql, ...)

db_data_type.OsqueryConnection(con, fields, ...)

sql_escape_ident.OsqueryConnection(con, x)

db_desc.OsqueryConnection(x)
}
\description{
Some small helpers to deal with the DBI<->dbplyr bridge
}
\keyword{internal}
