% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osquery-driver.R
\docType{methods}
\name{dbSendQuery,OsqueryConnection-method}
\alias{dbSendQuery,OsqueryConnection-method}
\title{Send a query to Osquery}
\usage{
\S4method{dbSendQuery}{OsqueryConnection}(conn, statement, ...)
}
\description{
Send a query to Osquery
}
\examples{
# This is another good place to put examples
}
