% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose.R
\name{osq_expose_tables}
\alias{osq_expose_tables}
\alias{osq_load_tables}
\title{Expose all (or selected) local or remote osquery tables as global environment dbplyr tibbles
or as a named list.}
\usage{
osq_expose_tables(tables = ".*", prefix = NULL, host = NULL,
  keyfile = NULL, quiet = TRUE)

osq_load_tables(tables = ".*", prefix = NULL, host = NULL,
  keyfile = NULL, quiet = TRUE)
}
\arguments{
\item{tables}{defaults to \code{.*} or all tables. Use a valid regular expression to match table names.}

\item{prefix}{if not \code{NULL}, each table in \code{tables} will be prefixed with this string.}

\item{host, keyfile}{same as for \code{\link[=dbConnect]{dbConnect()}}}

\item{quiet}{if \code{TRUE}, display a message with each table being exposed to the global environment.}

\item{tables}{defaults to \code{.*} or all tables. Use a valid regular expression to match table names.}

\item{prefix}{if not \code{NULL}, each table in \code{tables} will be prefixed with this string.}

\item{host, keyfile}{same as for \code{\link[=dbConnect]{dbConnect()}}}

\item{quiet}{if \code{TRUE}, display a message with each table being wired up}
}
\description{
This is intended as a shortcut for building a connection and wiring up to \code{src_dbi()}.
Exposing all of the tables is not recommended unless you're performing a large number
of diverse queries.
}
\details{
\code{\link[=osq_load_tables]{osq_load_tables()}} is a tad safer than \code{\link[=osq_expose_tables]{osq_expose_tables()}} since it does not
pollute the global environment and you can use a variable name for host
identification vs a prefix (though prefixes are still supported in \code{\link[=osq_load_tables]{osq_load_tables()}}).
}
