% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osquery-driver.R
\docType{methods}
\name{dbUnloadDriver,OsqueryDriver-method}
\alias{dbUnloadDriver,OsqueryDriver-method}
\alias{show,OsqueryDriver-method}
\title{Osquery class}
\usage{
\S4method{dbUnloadDriver}{OsqueryDriver}(drv, ...)

\S4method{show}{OsqueryDriver}(object)
}
\description{
Osquery class

Osquery show
}
