% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osquery-driver.R
\docType{methods}
\name{dbConnect,OsqueryDriver-method}
\alias{dbConnect,OsqueryDriver-method}
\alias{dbDisconnect,OsqueryConnection-method}
\title{Osquery dbConnect}
\usage{
\S4method{dbConnect}{OsqueryDriver}(drv, ...)

\S4method{dbDisconnect}{OsqueryConnection}(conn, ...)
}
\arguments{
\item{drv}{An object created by \code{Osquery()}}

\item{conn}{An object created by \code{OsqueryConnection()}}
}
\description{
Osquery dbConnect

Osquery dbDisconnect
}
\examples{
\dontrun{
db <- dbConnect(osqueryr::Osquery())
}
\dontrun{
db <- dbConnect(osqueryr::Osquery())
}
}
