% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osquery-driver.R
\docType{class}
\name{OsqueryConnection-class}
\alias{OsqueryConnection-class}
\alias{dbIsValid,OsqueryConnection-method}
\alias{dbListFields,OsqueryConnection,character-method}
\alias{dbListTables,OsqueryConnection-method}
\title{Osquery connection class.}
\usage{
\S4method{dbIsValid}{OsqueryConnection}(dbObj, ...)

\S4method{dbListFields}{OsqueryConnection,character}(conn, name, ...)

\S4method{dbListTables}{OsqueryConnection}(conn, ...)
}
\description{
Osquery connection class.
}
\keyword{internal}
