% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osqueryr-package.R
\docType{package}
\name{osqueryr}
\alias{osqueryr}
\alias{osqueryr-package}
\title{'osquery' Interface for R}
\description{
'osquery' \url{https://osquery.readthedocs.io/en/stable/} is an operating
system instrumentation framework for 'Windows', 'OS X (macOS)', 'Linux', and
'FreeBSD'. The tools make low-level operating system analytics and monitoring
both performant and intuitive. A full 'dbplyr'-compliant 'DBI'-driver
interface is provided facilitating intuitive and tidy analytic idioms.
}
\author{
Bob Rudis (bob@rud.is)
}
