% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osquery-driver.R
\docType{class}
\name{OsqueryResult-class}
\alias{OsqueryResult-class}
\alias{dbGetStatement,OsqueryResult-method}
\title{Osquery results class.}
\usage{
\S4method{dbGetStatement}{OsqueryResult}(res, ...)
}
\description{
Osquery results class.

Statement
}
\keyword{internal}
